requirements.txt
pyglet==1.5.21
cocos2d==0.6.9
